+++ moiza-domain/src/main/java/com/moiza/moizaspringbootserver/auth/spi/QueryTokenRefreshSpi.java
@@ -1,6 +1,6 @@
 package com.moiza.moizaspringbootserver.auth.spi;
 
-import com.moiza.moizaspringbootserver.auth.RefreshToken;
+import com.moiza.moizaspringbootserver.auth.domain.RefreshToken;
 
 public interface QueryTokenRefreshSpi {
     RefreshToken queryRefreshTokenByToken(String token);
